+++ ros1/src/data/march_realsense_reader/include/pointcloud_processor/parameter_determiner.h
@@ -136,23 +136,23 @@ protected:
 
     // Fill the foot locations to try cloud with a line of points from (start,
     // 0) to (end, 0)
-    bool fillOptionalFootLocationCloud(double start, double end);
+    bool fillOptionalFootLocationCloud(float start, float end);
 
     // Read all relevant parameters from the parameter yaml file
     void readYaml();
     int hull_dimension;
     int number_of_optional_foot_locations;
-    double min_x_stairs;
-    double max_x_stairs;
-    double min_z_stairs;
-    double max_z_stairs;
-    double y_location;
+    float min_x_stairs;
+    float max_x_stairs;
+    float min_z_stairs;
+    float max_z_stairs;
+    float y_location;
     double x_flat;
     double z_flat;
     double x_steep;
     double z_steep;
-    double min_search_area;
-    double max_search_area;
+    float min_search_area;
+    float max_search_area;
     double max_distance_to_line;
     bool general_most_desirable_location_is_mid;
     bool general_most_desirable_location_is_small;
